tests: include common in the right place
`pkgdep_toolpath` happens in scripts/pkgdep.sh
that changes export PATH

but the include of `scripts/common.sh` in `autotest_setup.sh`
happens before, so the path is not updated.

we did not see errors so far because accidentaly all
packages were already in the path, but if we remove one,
we can easily see the problem.

Change-Id: I74b7b18fcbd8630990692176e51062a357c1bac3
Signed-off-by: Boris Glimcher <Boris.Glimcher@emc.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/27289
Reviewed-by: Tomasz Zawadzki <tomasz@tzawadzki.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Michal Berger <michal.berger@nutanix.com>
Reviewed-by: Konrad Sztyber <ksztyber@nvidia.com>