test/common: fix CPU busy check when AIO is a backend device
If AIO is a backend block device then we're in the kernel realm in terms
of task status and 'uninterruptible sleep' is a valid state for the
SPDK process to be in. AIO is used for interrupt mode testing.

Fix busy CPU check by treating it the same way as R (Running).
Long term fix, as suggested in below comment, would be to track TSC in
reactor struct, however it requires more work. This one should resolve
the issue short-term.

https://review.spdk.io/c/spdk/spdk/+/26187/comment/1ef59015_332894ec/

Fixes #3634

Change-Id: Id1d8b2f0b1a45187bfa2e95edfc12511534b5bc8
Signed-off-by: Krzysztof Goreczny <krzysztof.goreczny@dell.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/27326
Reviewed-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Reviewed-by: Tomasz Zawadzki <tomasz@tzawadzki.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>