python: use system Python for make uninstall
By default "install" and "uninstall" actions don't
operate in virtual env.
"up pip uninstall" does not create a venv itself and
expects on to be already present before uninstalling,
which results in "No virtual environment found" error.

Check if we're running in a vritual env. If not then
add "--system" flag so that system Python is used
instead.

Fixes #3799

Change-Id: Ie4d7f34efcbaa5f5694637d92349d404b14186ee
Signed-off-by: Karol Latecki <karol.latecki@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/27363
Reviewed-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Tomasz Zawadzki <tomasz@tzawadzki.com>