dpdkbuild: remove use of findstring for CONFIG vars
No need to use findstring(), instead of just comparing the actual CONFIG var.

Unlikely to be an issue, findstring() matches any part of the var.
So any of the following would be interpreted as enabled: 'y', 'yes', 'nay'.

While here, fixed the 'or' for unit tests too due to the same reason.

Change-Id: Ib281cf060738bc968300301f188df95808e68e2f
Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/27380
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Reviewed-by: Konrad Sztyber <ksztyber@nvidia.com>