dpdkbuild: only build DPDK libs if OS supports it
rte_power and rte_vhost only work on Linux, when building DPDK on FreeBSD
following message could be seen:
power:	only supported on Linux
Resulting in both libraries being disabled.

Currently we are requesting to disable all libraries with exception of the ones
used by SPDK. In this case we do not disable power/vhost, so DPDK attempts to
build them. Yet does not treat it as failure when they are not built,
for example due to lack of OS support.

Next patch in series will be switching from disable_libs to enable_libs DPDK
meson option, as such any failure to build a requested library will be treated
as failure. To get ahead of it, do not request building those libs on OS other
than Linux.

Used CONFIG_VHOST, since it is already OS dependent - see configure.

Change-Id: Ie1bdc65dcc2881fd664cfc52e8e807fe7b431feb
Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/27394
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Reviewed-by: Mateusz Kozlowski <mateusz.kozlowski@solidigm.com>